:sparkles: 인가 로직을 Entity에 위임 (#63)
## 요약
PermissionChecker를 삭제하고 Entity에 구체적인 권한 체크 로직을 위임후 체크에 따른 처리는 Service 에서
진행하게 변경했습니다.

Service 안에 내부 함수로 만들면 Object로 통일한거랑 똑같은 거라고 보이긴 하는데 권한 체크에 실패했을 때의 인가
처리는 서비스마다 달라질 수 있어 서비스는 분리하되 현재는 정책대로 통일했습니다.

추가로 예외처리 발생을 만들어져있는줄 까먹었던 AccessDeniedException 으로 통일했습니다.

Recipe는 권한 체크가 없길래 일단 Entity에 메소드만 작성했습니다.

## 작업 사항
- PermissionChecker 제거
- UserPrincipal 정보 조회 편하게 멤버 추가
- 기존 인가 부분에 해당하는 로직 각 Entity 로 위임
- 각 서비스에서 공통된 인가 처리를 담당하는 메소드 추가
---
### 해결한 이슈
closes: #60

+++ src/main/kotlin/team/sparta/onehouronemeal/domain/comment/service/v1/CommentService.kt
@@ -3,7 +3,6 @@ package team.sparta.onehouronemeal.domain.comment.service.v1
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import team.sparta.onehouronemeal.domain.auth.common.PermissionChecker
 import team.sparta.onehouronemeal.domain.comment.dto.v1.CommentResponse
 import team.sparta.onehouronemeal.domain.comment.dto.v1.CreateCommentRequest
 import team.sparta.onehouronemeal.domain.comment.dto.v1.UpdateCommentRequest
@@ -11,6 +10,7 @@ import team.sparta.onehouronemeal.domain.comment.model.v1.Comment
 import team.sparta.onehouronemeal.domain.comment.repository.v1.CommentRepository
 import team.sparta.onehouronemeal.domain.course.repository.v1.CourseRepository
 import team.sparta.onehouronemeal.domain.user.repository.v1.UserRepository
+import team.sparta.onehouronemeal.exception.AccessDeniedException
 import team.sparta.onehouronemeal.exception.ModelNotFoundException
 import team.sparta.onehouronemeal.infra.security.UserPrincipal
 
@@ -43,7 +43,7 @@ class CommentService(
     fun updateComment(principal: UserPrincipal, commentId: Long, request: UpdateCommentRequest): CommentResponse {
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
-        PermissionChecker.check(comment.user.id!!, principal)
+        checkPermission(comment, principal)
 
         comment.updateComment(request.content)
 
@@ -54,7 +54,7 @@ class CommentService(
     fun deleteComment(principal: UserPrincipal, commentId: Long) {
         val comment = commentRepository.findByIdOrNull(commentId) ?: throw ModelNotFoundException("Comment", commentId)
 
-        PermissionChecker.check(comment.user.id!!, principal)
+        checkPermission(comment, principal)
 
         commentRepository.delete(comment)
     }
@@ -67,5 +67,14 @@ class CommentService(
         // 본인 댓글에 신고 안됨
 
         TODO()
+    }
+
+    private fun checkPermission(comment: Comment, principal: UserPrincipal) {
+        check(
+            comment.checkPermission(
+                principal.id,
+                principal.role
+            )
+        ) { throw AccessDeniedException("You do not own this comment") }
     }
 }
